requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
reportlab==4.0.9
requests==2.31.0
rpds-py==0.17.1
simplejson==3.19.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
